# openapi.yaml
openapi: 3.0.0
info:
  title: Project Spark Blueprint API
  version: v1
  description: API for Project Spark Blueprint application
servers:
  - url: http://localhost:5000/v1 # Assuming API v1 prefix is handled by Flask app
    description: Local development server
paths:
  /auth/register:
    post:
      summary: Register a new user
      operationId: registerUser
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201': # Flask-RESTX typically returns 201 for successful POST creating a resource
          description: User registered successfully. Returns token and user info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse' # Backend sends LoginResponse on successful registration
        '400':
          description: Invalid input (e.g., validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Log in an existing user
      operationId: loginUser
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid input (e.g., validation error for email/password format)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (e.g., invalid credentials)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Refresh an access token
      operationId: refreshToken
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Access token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          description: Unauthorized (e.g., invalid or expired refresh token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # New Project and Member Endpoints
...
components:
  schemas:
    UserBase: # Corresponds to UserBase in Pydantic
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        language:
          type: string
          enum: ['en', 'es']
          default: 'en'
          nullable: true
        role:
          type: string
          enum: ['admin', 'member']
          default: 'member'
          nullable: true
    UserPublic: # Corresponds to UserPublic in Pydantic
      allOf:
        - $ref: '#/components/schemas/UserBase'
      type: object
      properties:
        id:
          type: string
          format: uuid # Assuming UUIDs are represented as strings
        created_at:
          type: string
          format: date-time
    RegisterRequest: # Corresponds to RegisterRequest in Pydantic
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest: # Corresponds to LoginRequest in Pydantic
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    Token: # Corresponds to Token in Pydantic
      type: object
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
          nullable: true
        token_type:
          type: string
          default: 'bearer'
    LoginResponse: # Corresponds to LoginResponse in Pydantic
      allOf:
        - $ref: '#/components/schemas/Token'
      type: object
      required: # user is part of LoginResponse
        - user
      properties:
        user:
          $ref: '#/components/schemas/UserPublic'
    RefreshTokenRequest: # Corresponds to RefreshTokenRequest in Pydantic
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    RefreshTokenResponse: # Corresponds to RefreshTokenResponse in Pydantic
      type: object
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: 'bearer'
    ErrorResponse: # Generic error response
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Adding common responses to avoid repetition
...
security:
  - bearerAuth: [] 
# The global security definition means all endpoints require bearerAuth by default.
# For public endpoints like /auth/login and /auth/register, 
# you need to override this at the operation level with `security: []`
# Let's adjust paths for /auth/login and /auth/register to explicitly have no security

# Adjusting auth paths to have no global security
# This should be done by finding these paths in the existing openapi.yaml and adding 'security: []'
# For this edit, I will assume the edit tool handles merging this correctly.
# If not, a separate edit focusing only on auth paths might be needed.

# Path adjustments for auth (example, needs to be merged with existing auth paths)
# /auth/register:
#   post:
#     security: [] 
#     # ... rest of the definition
# /auth/login:
#   post:
#     security: []
#     # ... rest of the definition
# /auth/refresh: # Refresh token might or might not need Bearer auth, depending on flow.
               # Assuming refresh token itself is the auth for this one, also no Bearer.
#   post:
#     security: [] 
#     # ... rest of the definition 
...